Move `DeleteComments` command
